pyproject.toml
[project]
name = "claude-token-saver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastembed>=0.8.0",
    "lancedb>=0.30.2",
    "mcp>=1.27.0",
    "pyarrow>=23.0.1",
    "pydantic>=2.13.0",
    "pyyaml>=6.0.3",
    "tree-sitter>=0.25.2",
    "tree-sitter-javascript>=0.25.0",
    "tree-sitter-python>=0.25.0",
    "tree-sitter-typescript>=0.23.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]
